+++ src/tr2/global/const.h
@@ -48,7 +48,6 @@
 #define MAX_STATIC_OBJECTS 50
 #define MAX_ITEMS 256
 #define MAX_EFFECTS 100
-#define MAX_PICKUPS 12
 #define MAX_LEVELS 24
 #define MAX_LEVEL_NAME_SIZE 50 // TODO: get rid of this limit
 #define MAX_DEMO_FILES MAX_LEVELS